fedex-integration/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.0 ; python_version >= "3.12" and python_version < "4.0"
apscheduler==3.11.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
fastapi==0.115.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
holidays==0.60 ; python_version >= "3.12" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.1 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pgeocode==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
pycryptodome==3.20.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.5.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
spylib==0.10.3 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.38.6 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
tzlocal==5.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
